refactor: begin to refactor ACME certificate generation into the state pattern
